requirements.txt
django
djangorestframework
python-igraph
osmnx
networkx
scikit-learn
pathos
tqdm